Fix unlink calls for Redis Handlers

+++ packages/cache-handler/src/handlers/redis-stack.ts
@@ -120,6 +120,10 @@ export default async function createHandler({
                 from += querySize;
             }
 
+            if (keysToDelete.length === 0) {
+                return;
+            }
+
             const options = getTimeoutRedisCommandOptions(timeoutMs);
 
             await client.unlink(options, keysToDelete);
